Removed all code protections and encryptions from the source code
SOA is a complete  school management application written in pure PHP.

My vision for SOA is to make things easier for schools around the world.  What better way to do that than to make the source code open for every one to access, use, contribute and improve.

I hope you will find this helpful and useful.